+++ openghg/store/_eulerian_model.py
@@ -1,9 +1,8 @@
 from __future__ import annotations
 from pathlib import Path
-from typing import Any, DefaultDict, Dict, Optional, Union
+from typing import Any, Dict, Optional, Union
 import logging
 from openghg.store.base import BaseStore
-from xarray import Dataset
 
 logger = logging.getLogger("openghg.store")
 logger.setLevel(logging.DEBUG)  # Have to set level for logger as well as handler
@@ -42,7 +41,8 @@ class EulerianModel(BaseStore):
         compressor: Optional[Any] = None,
         filters: Optional[Any] = None,
         chunks: Optional[Dict] = None,
-        optional_metadata: Optional[Dict] = None,
+        info: Optional[Dict] = None,
+        **kwargs: Any,
     ) -> Dict:
         """Read Eulerian model output
 
@@ -74,12 +74,10 @@ class EulerianModel(BaseStore):
                 for example {"time": 100}. If None then a chunking schema will be set automatically by OpenGHG.
                 See documentation for guidance on chunking: https://docs.openghg.org/tutorials/local/Adding_data/Adding_ancillary_data.html#chunking.
                 To disable chunking pass in an empty dictionary.
-            optional_metadata: Allows to pass in additional tags to distinguish added data. e.g {"project":"paris", "baseline":"Intem"}
+            info: Allows to pass in additional tags to distinguish added data. e.g {"project":"paris", "baseline":"Intem"}
         """
         # TODO: As written, this currently includes some light assumptions that we're dealing with GEOSChem SpeciesConc format.
         # May need to split out into multiple modules (like with ObsSurface) or into separate retrieve functions as needed.
-
-        from collections import defaultdict
         from openghg.util import (
             clean_string,
             timestamp_now,
@@ -193,22 +191,11 @@ class EulerianModel(BaseStore):
 
             key = "_".join((model, species, date))
 
-            model_data: DefaultDict[str, Dict[str, Union[Dict, Dataset]]] = defaultdict(dict)
-            model_data[key]["data"] = em_data
-            model_data[key]["metadata"] = metadata
+            model_data = {key: {"data": em_data, "metadata": metadata}}
 
-            required = ("model", "species", "date")
+            lookup_keys = self.get_lookup_keys(optional_metadata=kwargs)
 
-            if optional_metadata:
-                common_keys = set(required) & set(optional_metadata.keys())
-
-                if common_keys:
-                    raise ValueError(
-                        f"The following optional metadata keys are already present in required keys: {', '.join(common_keys)}"
-                    )
-                else:
-                    for key, parsed_data in model_data.items():
-                        parsed_data["metadata"].update(optional_metadata)
+            model_data = self._add_additional_metadata(data=model_data, kwargs_metadata=kwargs, info=info)
 
             data_type = "eulerian_model"
             datasource_uuids = self.assign_data(
@@ -216,7 +203,7 @@ class EulerianModel(BaseStore):
                 if_exists=if_exists,
                 new_version=new_version,
                 data_type=data_type,
-                required_keys=required,
+                required_keys=lookup_keys,
                 compressor=compressor,
                 filters=filters,
             )