+++ openghg/transform/_transform.py
@@ -6,7 +6,7 @@ data to the object store via transformations.
 from __future__ import annotations
 
 from pathlib import Path
-from typing import Optional, Union, Any
+from typing import Any, Optional, Union
 
 from openghg.objectstore import get_writable_bucket
 from openghg.store import get_data_class
@@ -17,6 +17,7 @@ def transform_flux_data(
     database: str,
     overwrite: bool = False,
     store: Optional[str] = None,
+    info: Optional[dict] = None,
     **kwargs: Any,
 ) -> dict:
     """Read and transform a flux / emissions database. This will find the appropriate
@@ -33,14 +34,19 @@ def transform_flux_data(
         overwrite: Should this data overwrite currently stored data
             which matches.
         store: name of object store to write data to.
+        info: dictionary of tags for searching or other additional information. This info is *not* used to distinguish between datasources.
+            e.g. {"project": "paris", "comments": "Flat prior based on EDGAR totals by country"}
         **kwargs: Inputs for underlying parser function for the database.
-            Necessary inputs will depend on the database being parsed.
+            Necessary inputs will depend on the database being parsed. Any
 
     Returns:
+        dict containing datasource UUIDs data is assigned to
     """
     bucket = get_writable_bucket(name=store)
     dclass = get_data_class("flux")
 
     with dclass(bucket) as dc:
-        result = dc.transform_data(datapath=datapath, database=database, overwrite=overwrite, **kwargs)
+        result = dc.transform_data(
+            datapath=datapath, database=database, overwrite=overwrite, info=info, **kwargs
+        )
     return result