+++ tests/store/test_flux.py
@@ -1,6 +1,7 @@
 import pytest
 import os
 from helpers import clear_test_stores, get_flux_datapath
+from openghg.objectstore import get_metakey_defaults
 from openghg.retrieve import search, search_flux
 from openghg.store import Flux
 from openghg.standardise import standardise_flux, standardise_from_binary_data
@@ -9,6 +10,7 @@ from openghg.util import hash_bytes
 from pandas import Timestamp
 from xarray import open_dataset
 from typing import Any, Union
+from unittest.mock import patch
 
 
 @pytest.fixture
@@ -449,9 +451,9 @@ def test_flux_schema():
     # TODO: Could also add checks for dims and dtypes?
 
 
-def test_optional_metadata_raise_error(clear_stores):
+def test_info_raise_error(clear_stores):
     """
-    Test to verify required keys present in optional metadata supplied as dictionary raise ValueError
+    Test to verify required keys present in info dictionary raises ValueError
     """
     folder = "v6.0_CH4"
     test_datapath = get_flux_datapath(f"EDGAR/yearly/{folder}")
@@ -460,10 +462,16 @@ def test_optional_metadata_raise_error(clear_stores):
     date = "2015"
 
     with pytest.raises(ValueError):
-        proc_results = transform_flux_data(store="user", datapath=test_datapath, database=database, date=date, optional_metadata={"domain":"openghg_tests"})
+        proc_results = transform_flux_data(
+            store="user",
+            datapath=test_datapath,
+            database=database,
+            date=date,
+            info={"domain": "openghg_tests"},
+        )
 
 
-def test_optional_metadata():
+def test_info():
     """
     Test to verify optional metadata supplied as dictionary gets stored as metadata
     """
@@ -473,7 +481,13 @@ def test_optional_metadata():
     database = "EDGAR"
     date = "2015"
 
-    proc_results = transform_flux_data(store="user", datapath=test_datapath, database=database, date=date, optional_metadata={"project":"openghg_tests", "tag":"tests"})
+    proc_results = transform_flux_data(
+        store="user",
+        datapath=test_datapath,
+        database=database,
+        date=date,
+        info={"project": "openghg_tests", "tag": "tests"},
+    )
 
     version = "v6.0"
     species = "ch4"