+++ tests/store/test_obscolumn.py
@@ -47,7 +47,7 @@ def test_read_openghg_format():
         assert np.isclose(ch4_data["xch4"][0], 1844.2019)
 
 
-def test_optional_metadata_raise_error():
+def test_info_raise_error():
     """
     Test to verify required keys present in optional metadata supplied as dictionary raise ValueError
     """
@@ -68,11 +68,11 @@ def test_optional_metadata_raise_error():
             satellite=satellite,
             domain=domain,
             species=species,
-            optional_metadata={"domain":"openghg_test"}
+            info={"domain":"openghg_test"}
      )
 
 
-def test_optional_metadata():
+def test_info():
     """
     Test to verify required keys present in optional metadata supplied as dictionary raise ValueError
     """
@@ -90,7 +90,7 @@ def test_optional_metadata():
         satellite=satellite,
         domain=domain,
         species=species,
-        optional_metadata={"project":"openghg_test"}
+        info={"project":"openghg_test"}
     )
     col_data = search_column(species="ch4").retrieve_all()
     metadata = col_data.metadata